backend/requirements.txt
# En-Shirube System — Backend Dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.14.0

# Validation & settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI / LLM
openai>=1.50.0
anthropic>=0.40.0

# PDF generation
weasyprint>=62.0

# Utilities
httpx>=0.27.0
